backend/requirements.txt
aiohttp==3.8.1
aiohttp-devtools==1.0.post0
aiosignal==1.2.0
anyio==3.6.1
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
bidict==0.22.0
charset-normalizer==2.1.0
click==8.1.3
devtools==0.9.0
environs==9.5.0
executing==0.10.0
frozenlist==1.3.1
idna==3.3
marshmallow==3.17.0
multidict==6.0.2
packaging==21.3
Pygments==2.13.0
pyparsing==3.0.9
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.1
six==1.16.0
sniffio==1.2.0
watchgod==0.8.2
yarl==1.8.1
